tests.workflows: add conftest for common fixtures
